Renaming legals.dart
