+++ lib/routes.dart
@@ -16,7 +16,7 @@
 */
 
 import 'package:flutter/cupertino.dart';
-import 'package:picos/screens/legals.dart';
+import 'package:picos/screens/legals_screen.dart';
 import 'package:picos/screens/my_medications_screen/add_medication_screen.dart';
 import 'package:picos/screens/my_therapies_screen/add_therapy_screen.dart';
 import 'package:picos/screens/my_therapies_screen/my_therapies_screen.dart';
