+++ lib/screens/legals_screen.dart
